+++ awc/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - 2023-xx-xx
 
+- Add `awc::Connector::rustls_021()` method for Rustls v0.21 support behind new `rustls-0_21` crate feature.
+- Add `rustls-0_20` crate feature, which the existing `rustls` feature now aliases.
 - Minimum supported Rust version (MSRV) is now 1.68 due to transitive `time` dependency.
 
 ## 3.1.1 - 2023-02-26